fix(web): защита FormField от отсутствующего ребёнка
